fix; erroneous 'pwd' prints removed
The statement using subprocess.run to query the status of the git repos
included 2 calls tp "pwd". The output of these calls was added to the
content of the stdout element in the result, which was printed when
attempting to print the hash.
I removed the 2 calls to "pwd" from the 2 subprocess.run statements and
now it seems to work fine.

I also again (?) removed the "capture_output" arguments to
subprocess.run which doesn't exist yet for python 3.6 (I think).