feature: iPCR-trim now trims 3prime
In case gDNA is very short reads may read into the downstream adapter
sequence. The script now checks whether the reverse complement adapter
sequence is present in the read in if so trims the adapter and
subsequent sequence from the reads. This only applies to a few percent
of the reads.