fix bam2annotBed: SNP overlaps deletion in read
If genomic positions are deleted from a read the genomic positions no
lonlonger correspong to any position in the read. Now a SNP position is
checked against genomic positions actually corresponding to positions
in the read ('r.query_sequence.index(str(snp.start))'). Only if the
position exists the base identity and other info is inferred. If the
positions is deleted from the read 'None' is returned for base identity