feature; new script: merge iPCR, cDNA
script: code/merge_iPCR_cDNA.R
new rules in SuRE-snakemake

The script is a simple R script which reads the iPCR bedpe file and
merges it with the counts in all cDNA trimmed-table-count files

The rules in SuRE-snakemake are a straightforward rule
('merge_ipcr_cdna') which calls the R script with proper arguments.
In addition a new target rule is created ('merged_ipcr_cdna').