fix bam2annotbed: log extensive msgs 2 local log
If SNPs are discarded because they are neither simple SNP nor INDEL
they are printed to stdout. This resulted in very large log files at
the pipeline level.
Now the script takes a log file and writes the extensive log entries
(thethe annotation of discarde SNPs) to this separate log file.