feat bam2annotBed; test len(GT) fixes haplo genome error
If GT is haploid (eg on chrX) calculating iupac code results in error
as it takes two bases as input.
Now I first test the len(GT); if len=1 this means the position is
haploid and instead of calculating iupac I directly assign the
base-identity at that position. If len(GT)!=1 the string (eg '2|0')
is split and the two values are used to calculate iupac.
A similar code change is done in function 'annotate_snp_in_read(..)'