fix bam2annotBed; don't 'r.end-1' for find snp
Previously I did '-1' from the end of the SuRE-fragment when looking
for overlap with SNPs. I knew this was incorrect but I did it because
the results were incorrect without it. These incorrect results were
probaprobably the result of other errors in the coord-system usage
(which I now hopefully fixed).
Now the substraction no longer happens...