fix BC2bedpe; trim readID in BC import
Some time previously there was (and still is???) an issue with readID's
containing white spaces, which were differently processed in bedpe's
and info files. The solution is to trim the readID here, to make them
comcompatible.
Now I found that the trimming was erroneous if there is NO wihite space
in the readID. So I replaced the reg expr from "r'^(.*) .*$'" to
"r'^(\S+).*$'".