feat iPCR-map; bowtie  --sam-no-qname-trunc
This argument refrains bowtie from trimming the read name. Now there
will be no longer a clash between the read names used in the bowtie
output (bam, bedpe files) and the info file generated by cutadapt.
There still is an issue with trailing '/1,/2' though (which needs to be
fixed in addBC2bedpe)