fix: import info_file now skips unparsed lines
Previously if info file was imported (for extracing barcode sequences)
it could fail if the first line contained a parsed read for which the
parparsing failed. In that case there is not a complete set of fields
in the line. Subsequent lines were not imported correctly; all barcode
sequences were set to NaN.
Now the info file is first read line by line to find the first line
contacontaining a correctly parsed read. This line number is then used
as first line to import the file into a pandas data frame. This way the
5th column is read correctly an if subsequent lines contain unparsed
readreads this does not distort the data frame.